Registration Number algorithm update
City Health Office users requested to change how the system provides registration numbers. When renewing, instead of retaining a Yellow Health Certificate's registration number for the current year, it will be issued a new one. The registration number algorithm is now dependent on the number of the latest registration number and not on the total count of rows queried. It is also ensured that the new algorithm is adaptive to other controllers utilizing the function.